Grenades:  * Fixed it so grenade fire classes are properly using the DH ammo classes  * Added F1 grenade ammo class  * Changed grenades to have 3 max and 1 initial
Notes:
This is a pretty significant change to the game. It will bring a lot of value to the new construction (grenade boxes). In Push, with no grenade box, players will have less grenades, or will they?

Sure you only spawn with 1 grenade now, but you can pick up 2 extra. You can almost always find grenades on the ground in Push, its pretty predictable where casualties will be. So though you spawn with only 1, if you loot you will have 3 most likely.

In a way I'm excited to see how much of an effect this will have on the game, both Push and Advance. I think it will result in less overall grenade spam, but allow for the occasional large nade spam. So it should even out.

Good luck grenade throwing, hope to be out of range from you on the battlefield!