Changed the text drawing method used to draw squad-mate numbers on the situation map and the numbered seats on the vehicle clock image. This allows for more flexibility, since it uses actual native text rendering instead of the texture-based approach used earlier.
This was mainly done to accommodate a change that would make it easier to identify the assistant on the map by having his icon be paired with an "A", instead of his squad member index number. In addition, the squad leader's player icon text is now "SL" instead of "1", finally making the situation map consistent with all other parts of the interface.