Adjusted Axis AT guns to be able to aim lower. These values are not meant to be realistic, but allow for more realistic aiming. No one in their right mind would setup a Pak40 to not be able to aim low enough to hit anything. The AT guns fall backwards because they lack realistic sandbag support under the gun's support legs, their general angle is higher than should be. The "unrealistic" angle is to adjust for this problem.
AKA this is mostly for Advance, as before it was the leveler's job to insure the AT gun placement wasn't bad. However, with players placing them, they will struggle more to get the AT gun to be able to aim low enough.

This will not RULE out the issue altogether though, but it will help to insure AT guns are more usable after being built.