Arbitrarily increased Panther's side armor values. Here is why:
The panther's side vulnerability is of course a fault of the tank. However, there are some factors that aren't taken into consideration regarding its side armor. Ignoring these factors has lead to some undesired outcomes of the tank's combat effectiveness. Currently, the Panther G (the better one) is easily killed by a Stuart/Greyhound from the side. Though it was possible for them to destroy a Panther from the side, it shouldn't be as easy as it is currently.

MattUK's huge refactor to upper/lower hull armor thickness and slope had a HUGE impact on the Panther. Basically, it opened up the tank's side armor to all sorts of "penetration" that wasn't possible before the change. Simply because before hits on the lower part calculated the upper slope. In a way, this is what MattUK wanted. However, the 5.1 tank penetration code was carefully "adjusted" to be how the original team wanted it to be.

This change attempts to put the Panther back to where it was before this upper/lower hull thickness/slop feature implementation and be more like 5.1.

Additional rationalization of the changes:
1) The PantherG side skirts are not used in the calculation against APCBC, which they should. However, in a way those skirts should help against all AP types just not as much as against HEAT and APCBC.
2) Treads, wheels, and other track components aren't considered on the lower hull side armor thickness. Those could actually account for something depending on the velocity, round type, etc.

Notes:
I opened up 5.1 and used the Stuart (37mm APCBC), trying to destroy the Panther on Target Range. I noticed it was extremely difficult as 90% of the time if the angle wasn't perfect (0 degrees) the round would shatter. Then I went in 8.0.8 and did the same thing and the Stuart was penetrating the Panther like it was nothing. After these changes, its much closer to 5.1, but still not perfect.