Raised the max pen ability of the 37mm back to what it was as I did find some pen tables of 71mm pen capabilities. This will allow it to pen. a Tiger tank on the very low rear if the angles are just right. Very likely to never happen, but it is possible.
This was done at request of @Saulniv with this evidence:
"i see on github that for the next pacht you will nerf the greyhound AP shell. but in the game it is already underperfoming. It can't penetrate the rear of the tiger at point blank range, which it can totally do IRL. i even have source !: http://www.dtic.mil/dtic/tr/fulltext/u2/a952910.pdf#page=23
page 12 of the doc is where the tiger wrecking happen"

Notes: This means the 37mm will have the potential it had, but the fall off of pen capabilities will be harsher. Meaning it will need to get closer to do what it could do before.