Significant changes to DHZoneInfo:  * Added ability for server to change its view distance and replicate it to clients so they can adjust their fog accordingly  * Added admin command to change server view distance for testing
Notes:
This is not finished, there is still a bug where a client connecting has FogDistanceMin instead of FogDistanceEnd, not sure why (attempted several fixes only one had promise, but broke everything else)

The idea of this is that it STILL supports the dynamic fog ratio, which is kinda amazing, though it needs more testing and perhaps a bit more understanding. I wanted to commit this as it is working for the most part and now its time to try lots of stuff to debug it.