Changes to supply rates:  * Decreased the bonus supply the main cache generated from 240 to 200  * Increased shared supply generated between all caches from 180 to 600
Notes:
So I knew that supply would be in much higher demand and that demand would carry on longer into the round than it did before. However, supply is now annoying to get as the majority of the supply is stuck at main requiring players to constantly drive back and forth (not fun).

I'm not lowering the main supply's bonus generation much, because I think if you goto main you should be certain to have some supply, that supply is also used for supply trucks that spawn and I think if we lower it too much, supply trucks will hardly ever spawn with much supply if your team has a lot of caches.

If your team has less than 4 caches, they will generate more than they did in v8.0.9, if you build 4 they will generate the same as they did in v8.0.9 and if you start building more than 4 they will begin to not generate as much.

We can adjust as needed, but I think these values will put the "supply" back onto the battlefield close to where it was at in v8.0.9 during a healthy amount of caches (less than 6).