PPSH41 update:  * Lowered recoil significantly  * Added select fire ability for single shot (required an ugly hack to do)  * Added sounds for single fire mode
Notes:
I've tested, no log warnings and it works in MP. Its not ideal to hack in a feature for this gun like so, especially because the PPD will want the same treatment. However, I'm not sure of a better way to do it without creating tons of bugs and rewriting nearly the entire FireMode system. For now this should work and at the least improve this weapon's behavior for players.