Munition System Update:  * Disabled munition drainage on Advance (so it is no longer being used in any gamemode, perhaps we will use it in a future one)  * Lowered default munition percentage from 75% to 60%  * Removed the score and death munition reward system
Notes:
Many players expressed how annoying it is to constantly be out of ammo late in the round. I have to agree, its not particularly fun to be out of ammo. The situation of low ammo is idealistic, not necessarily practical for a game. However, I am an idealist and therefore I've lowered the default munitions from 75% to 60% so ammo will be a bit more of a problem (but not too much of one). There will be times you run out, but its because you are a bad ass or one-itchy-finger-trigger-mf.