Modified reinforcement cost of vehicles:  * Reduced many costs by 2-3 and some by nearly 50%
Notes:
This is a good change because losing tanks impacts the fate of the battle TOO much. Its not particular fun or realistic to lose because only a couple players on the team were wasting tanks left and right. The punishment for losing tanks needs to reflect on that specific player, not the team. Its a shame to see matches lost when the team's infantry were doing amazing, even defending their early objectives well, but lose because a couple tankers threw tanks at the enemy non-stop.

Some calculations to consider:
Before a tanker could waste 11-26 reinforcements every minute (2 minutes to be more realistic)
Infantry can waste about 4 reinforcements every minute (2 minutes to be more realistic)
Now a tanker can waste about 7-15 reinforcements every minutes (2 minutes to be more realistic)

So as you can see, tanks are a costly thing. Ideally we need a system to either get rid of underperforming crewmen or punish them when they don't manage enough kills or don't survive for long enough in a tank. For now though reducing the negative impact they can have on a team is simple and quick.