Several changes and implementations:
DarkestHourGame:
 * Removed function from DarkestHourGame that is no longer used (because of this update)
 * Improved the CaptureObj() exec function in DarkestHourGame

"Improvements" to objectives:
 * Added a hash table for server to get objectives by tag very quickly (no looping the GRI.DHObjectives!)
 * Objectives can now be setup for requirements with objective tags (not just index)
 * Edited some categories of variables in DHObjective to make it easier to find the requirements to capture the objective

Notes:
So this is backwards compatible, it should not break any maps. Eventually I want to deprecate the index stuff, but only once we update the maps that use it. The downsides of this commit is that I had to replicate 2 bools for something that shouldn't or technically doesn't need to be replicated. However, I could not figure out a good/reliable/nonhack way to do it. So for now the replication will be acceptable. This should be working, tested in SP and MP under both new and old system.