Changes to the G43:  * Brought over the G43/G43Scoped from RO assets to DH, so we can edit the animation values  * Slightly adjusted the G43 weapon model FOV when sighted (very minor, but you can see a bit more of receiver)
G43 Changes:
 * Sped up the fire animation of the G43 significantly (more than double the animation rate)
 * Lowered the accuracy
 * Added a bit more horizontal recoil

G43 Scoped Changes:
 * Lowered the accuracy slightly (now more accurate than its un-scoped version)

Notes (Important):
We get a few complaints from a couple players regarding the G43 recoil, and complaints have been brought up in the past by many. Most players don't like the G43 for some reason, well that reason isn't recoil. It actually had less recoil than the M1 Garand, the problem was the fire animation being way too long. All other weapons have very short fire animations, but the G43 had an additional second or two to "realign" the sight. So basically it was impossible to shoot the gun effectively without waiting 2-3 seconds between shots, making it annoying to use. I've sped up the animation so players won't have this 2-3 seconds wait time to fire sighted. This means that the G43 just got a lot better at killing multiple enemies, as a good player can fire and kill multiple targets much more quickly.

With only the animation change, the G43 would out perform the M1 Garand, which as we know from history should not be the case. So I've raised the recoil to match the Garand, but I've also raised the spread to make it less accurate than the M1 Garand.

Overall I still think this is a huge buff to the G43, it will feel like a different weapon and I think players using it will appreciate the change.