FG42 changes:  * Adjusted recoil to behave more reasonable (harder to control when iron sighted (standing/crouching/prone) and also have a bit more recoil when deployed (it was a near laser before)  * Adjusted weapon FOV to have larger sights when aiming  * Should be able to be resupplied by other players now
Other changes:
 * Changed some default properties around (no functional changes)

Notes:
FG42 feels like a more polished weapon now for DH, meaning I think its closer to where it needs to be regarding effectiveness and behavior. This is basically an overall nerf to the FG42, but many agree it needed one. It is still a great choice though, really the only downside is how quickly the magazine runs empty. It is still like a shotgun in close quarters and is still amazing as a LMG. The biggest hit to it IMO is that it isn't the greatest "rifle" anymore. Its best as a LMG now. I think this is basically a half-way point between how the weapon behaved in 5.1 (like crap) and 8.0 (like laser).