Transport Vehicles Update (unarmored transport vehicles, half-tracks, and the bren/universal carrier):  * Are now much more vulnerable to HE rounds (including HTs, 75mm HE can now 1-shot destroy HTs)  * Are now more vulnerable to AP rounds (disabling them after just 1 hit no matter if its an engine hit or not)  * Wheeled vehicles are more vulnerable to generic body damage (before they could take 1000 damage before being disabled, now they can only take 200-300)  * Wheel damage slows vehicles down more (this includes HTs)  * Increased engine health, so unarmored vehicles can take a few more small arm rounds to the engine before being disabled
Notes:
This is an obvious nerf to unarmored transport vehicles, half-tracks, and the bren/universal carrier. These vehicles aren't meant to be used in ways they are currently being used. They can still survive an AP round, but are disabled and will likely need to be quickly abandoned. Direct hits with HE are now extremely effective against these vehicles and should basically 1-shot them the majority of the time. To give them a bit of survivability I've raised their engine health a bit so they can take a few more small arm rounds.