Updated the Game Settings page:  * Added hints to various settings to help players understand what they do or what we recommend  * Adjusted the "Connection" setting to instead read more accurately "NetSpeed" added hint and adjusted the values
Notes:
We raised the low end values of the NetSpeed because those settings are unplayable and were causing lots of issues for players who selected them. Players if they so desire can still achieve very low NetSpeeds by editing their INI file, but we don't recommend having less than 5000. We will look at adjusting NetSpeeds on the server end and I want to do some diagnostics regarding network traffic the servers use and if they are bottlenecking there, if not we could turn up the rates a bit more.