Made it so the spectate self mode is always valid. This is a wishful-shot-in-the-dark to fix the crash on death issue some players have. Slightly improved logic in another function.
Notes:
After both Basnett and I reviewed the spectator code and changes he made around the time the crashes started for players, we think its something to do with the spectate changes, but the only thing we see that could cause it is some strange hardware related issue where the rendering doesn't know how to handle a draw call at death, which might be related to the fact we made spectate self invalid if the player has no pawn. AKA for some very odd reason, some hardware need to be able to view self-spectate mode at death. As stated this is a shot-in-the-dark and we have no clue if it actually fixes the problem, if it does GREAT! If not the next step is to completely revert the spectator changes he made and see if that fixes them.