A couple of changes to Patron Rewards and Map Voting:
* Fix for https://github.com/DarklightGames/DarkestHour/issues/974

Note: The fix was forcing all the players to revote when the match ends. Meaning when the round ends, you resubmit your vote automatically, which updates the weight.

* Patreon rewards have been changed, they are now based on # of players on the server. This balances the power of patrons, so they have less additional votes with lower player counts, but higher additional votes with high player counts. I think this is a good change because, before the # of additional votes with 64 players was not very many, and with 10 players was way too much. I will continue to update them to make sure they are reasonable.

Note: Because the vote value is based on # of players on the server, if players begin to leave during the map vote, it will lower the patron vote power and it will not be accurate. So its possible once the vote timer ends, it selects a map different than was shown as most weight. This is only speculation, I need to actually test it in a multiplayer environment. If it is the case, then a simple fix is base the vote value on the # of players that were on the server when the map vote started. That or update all the votes everytime someone leaves.