Fixed a bug where inventory cycling with a radio in inventory using MouseWheelDown (PrevWeapon) actually called (NextWeapon) preventing from proper inventory cycling if radio existed. Somewhat mentioned in this report: https://github.com/DarklightGames/DarkestHour/issues/834
Notes: After spawning you still aren't able to cycle weapons until you drop and pick up the K98. I'm not sure what the cause of that bug, but I think it only occurs to SP, which leads me to think its a problem with Authority and/or timing of events.