Fixed https://github.com/DarklightGames/DarkestHour/issues/975
Notes:
The issue was it was generating AmmoMags, which are used by rockets differently. Had to remove a line that did that. That line was added to resolve this very old and minor bug:

https://bitbucket.org/darklightgames/darkesthour/issues/270/after-weapon-is-empty-drop-weapn-and-pick

Which I think is resolved in another way now, as I could not reproduce that bug after removing the line. That said we should keep and eye out in the logs for issues.