Update to server self-optimization system:  * Added a function in DHZoneInfo to change the server's fog distance based on a ratio value  * Added a function so the server can now make use of DHZoneInfo's `bUseDynamicFogDistance` and lower the view distance to improve performance for strikes  * Added a log if the server receives a performance strike, so server admin can know what levels are getting strikes with a log search
Notes: Finally my dream is realized! A server will now lower its view distance if performance suffers (assuming the level uses DHZoneInfos). The goal is to never have a server receive a strike, but if it does, there are multiple things that it can do to help ensure a second strike isn't received. I think I'm done for awhile now on server performance and the self-optimization system, just bug fixes if needed.