RPG43 update:  * Lowered amount from 3 to 2, this includes max amount carried  * Reduced damage from 200 to 180, this only affects splash damage to crewmen, passengers, and infantry  * Reduced impact damage from 250 to 200, this only affects direct hits to vehicles  * Increased damage multiplier from 2 to 3 for indirect throws or those that hit the top armor
Notes:
RPG43s are very gamey and I think no matter what we do to them they will be feel gamey to players. Part of their absurdity is the quick consecutive throwing, so dropping the amount from 3 to 2 will help reduce that by 33%. Second is how effective they are with tactless direct throwing. The grenades were very ineffective against side armor, like to the point they were worthless even to a Panther tank's side armor. Originally I designed them this way, where direct throwing was useless and an upward indirect throw with the grenade coming down was required for them to be effective. At some point changes completely flipped that around to the point the top hits were not able to even do any damage (which I think was really just a bug), but it also made the direct throwing pretty effective also. Later I fixed the top bug and applied damage bonuses for top hits, but I think that wasn't enough and they are still be used near-exclusively as directly thrown AT grenades. These changes hope to rectify that.

References:
"The most practical application would be to throw it high over a tank so that it will drop and strike the top of the turret or the engine compartment. Dropping it from a roof top would be another application."
http://www.inert-ord.net/russ02i/rpg43/index.html