Removed the exception of same-squad members into locked vehicles. Meaning you can lock out people in your squad.
Why?
Many times tankers want to be "included" in a good infantry squad. They might get reliable communication (good targets) and good security support, but suddenly Joe joins the squad and steals the tank and drives it into a minefield. That being the exact thing we implemented tank locking to prevent. So now players have the option to only let in others they want.

Notes:
We need to make it more evident how to lock/unlock your tank, etc.