Fixed https://github.com/DarklightGames/DarkestHour/issues/978
Notes: I noticed there were 2 different statics for the intact statics of the Opel Blitz, one with a higher poly count which had some artifact polys hidden in it. I set its texture to blue so we can detect it in levels and replace it, after we replaced all of them (aka don't notice anymore we can delete it).