Significant increases to the supply cost of guns. So now there is no way for a supply truck to create more than 1 gun of any type without needing to resupply first. The 88 will require more than 1 truck (2100 supply).
Notes:
Having a single truck create 3 (nearly 4) ZiS3 guns without needing to resupply, or 2 pak40s, or 2 AT57s, is a major problem for game play. Overall, the AT guns are fun to use/have, but they are hurting the game play significantly in the quantities I'm seeing. Attackers can struggle on levels if these AT guns are built up in large quantities. Increasing their cost specifically so trucks have to resupply before placing a second will slow down a team's ability to amass them. Players will have to be more careful where they place them.

The problem with this change, is it devalues the lighter AT guns at least in comparison to the heavier AT guns. However, I feel something needs to be done, because over saturation of even the smaller guns isn't helping game play its hurting.

In a way construction (especially AT guns and HQs) sorta keeps a winning team winning, as their build up of guns and supply continues, and the losing team's job gets harder and harder. This is a design choice we made from the beginning and we are fine with it to a point. Having 5 enemy AT guns fire HE at your approach is not fun, especially when you see them constructing a 6th and no one can do anything about it because of the other 5 guns.

If the over saturation continues regardless of this cost increase, then we will need to increase the duplication radius of them to higher values. For now this I think is at least a step in the right direction.