Removed the heavy engine damage thresholds to use the 0.5 base value
Significant reduction to the health of WillysJeep, Kubelwagen, and GAZ67:
 * Reduced the vehicle's health (before fire) from 400 to 100
 * Reduced the vehicle's engine health from ~25 to 10

Notes:
The reduction to the vehicle's health is actually more, because small arms do double damage now. Making the effective health change more like (400 to 50).
Before a MP40 required a little over 5 magazines to disable a Jeep, now it will take less than 1.

The reduction to the vehicle's engine health is also more significant.
Before a MP40 required 13 rounds to disable, now 3.
Before a K98 required 5 rounds to disable, now 1.

The problem is, in testing we almost always test the vehicle stationary, so its very easy to hit. However when these vehicles are used on a server with various pings and are moving, its extremely hard to him them. Making the 13 rounds or 5 rounds a lot harder to do than it sounds. These are rear echelon / support vehicles not battlefield clown cars, this will HOPEFULLY force players to use them properly/authentically and not risk them so haphazardly.