Deprecated classes inside DH_ATWeapons by having all the classes (but 1) extend DHDeprecated (a note class)
Also I created a new "DeprecatedPickups" class and have both "DH_PlaceablePanzerFaustPickup" classes extend it...

This does not mean they are even 50% deprecated as they will still spawn a DHPanzerfaust, later I plan to at least make them nonplaceable. The goal here is to make it so I can fix levels. Because the DH_ATWeapons version of "PlaceablePanzerFaustPickup" takes priority in the editor (even if I make it extend note) making it so I can't even place the proper "DH_Weapons.DH_PlaceablePanzerFaustPickup" version. And when I do use the proper "DH_Weapons.DH_PlaceablePanzerFaustPickup" the editor gets confused and thinks I did the DH_ATWeapons version. Gah!

I created a new (untouched) class called "DH_PlaceableFaustPickup" which is to be used instead and will eventually replace all the DH_PlaceablePanzerFaustPickup actors.

I also updated DH-Target_Range.rom as the first level to be DH_ATWeapons package free! I figure it was best to start with Target_Range as it had all sorts of references to DH_ATWeapons.

Notes:
So having 2 classes named the same thing is a BIG NO NO. I think the editor/engine gets confused and stores copies of the actor inside the level! I call this a ghost reference and the only way I know to get rid of the reference is to save myLevel package & copy+paste the actors to a fresh loaded new level.

It is very suspect that these ghost references are stacking up and act sorta like a memory leak of save data in the level.

The moral of the story is we shouldn't have done this same-name-cross-package-actor-reference-backwards-compatibility. Because now its infested a lot of levels and is likely doing harm to them (ghost references + ghost memory). And its also a bitch to remove.