Several Changes:  * Removed unused bUsingController variable in DHPlayer (was being set, but not actually referenced)  * Added a toggle run for infantry, added the setting to movement controls, no default bind (not a required control)  * Fixed a bug with the NetSpeed option, made it easier to make changes in the future by removing magic numbers as index references
Notes:
The toggle run works by pressing the key once, the user can cancel the auto run by hitting the key again OR moving with their normal movement keys once (hitting WASD). It is a bit of a hack implementation, for a couple reasons. I did have to use the client-sided PlayerMove() function, which is called in Tick(). However, it is only 1 simple if statement (on the client side) if the player is not auto-running.