Added a FPS limiter to DHMainMenuGame
Notes:
The system basically does more and more tick calculations if FPS is > Desired. Therefore it does not instantly limit FPS, it can take a second or so for it to bring the FPS down. The min/max FPS for the menu is 60-144, currently not changeable as it seems kinda silly to have options for that, but can easily add. This will hopefully stop coil whine on player's PSUs and/or GPUs, it does for me!