Hurtgenwald Fixed https://github.com/DarklightGames/DarkestHour/issues/1019
Notes:
So the issue was a BSP/Antiportal hybrid under the terrain. Either in the process of copy+pasting it, the geometry became invalid or some how it just magically worked before. The geometry was bad on it, because it would have invisible parts of its surface, which can and will slay players (even across) the map. The nice thing is those anti-portals are absolutely pointless, they weren't doing anything for FPS or the network, so their removal had no downsides (only upsides).

Other strange things:
I noticed after doing the ghost reference removal a third time. I noticed that the one of the terrain had a lot of invisible squares. I think for some reason when you copy+paste all actors, terrains with a lot of invisible sqaures do not copy correctly. You have to later go back and copy+paste that specific actor because otherwise the invisible parts are all screwed up. This is important to remember in the future as this ghost reference removal process can bug out terrains.

Help notes:
To easily fix this, I looked up the culprit commit (the one that causes the bug), then in the log, looking at the file for the commit, I right-click and goto save-as and to restore the work-tree select (before). This is super handy as I used it twice, once to replace the existing map and the second to make a "refernce" version. Which I used to realign the terrain locations and re copy-paste terrain actors etc.