Removed the debug wall from the net speed 10000 unlock
Notes:
Currently uses a magic number of 30000 as a max. For now lets keep max client rate 30000, it seems to be a good max value, as I was not able to actually use that much with 64 players. (max netspeed I noticed was 22000). If a server wants to lower it, it should be possible by changing MaxInternetClientRate in its config under category [IpDrv.TcpNetDriver]. If that is not true (more native bugs), then I will look into the matter (likely requiring another forced console command on the server).