Several changes:  * Implemented https://github.com/DarklightGames/DarkestHour/issues/1025  * Removed squad spawned vehicles (this isn't really useful, nearly everyone is in a squad and now the more important vehicles are controlled in a new way)
Notes:
So some functionality of the bMustBeInSquadToSpawn system was hijacked by the new bRequiresDriverLicense system. For instance, the error type used by bMustBeInSquadToSpawn for the deploy menu was replaced with  NoLicense instead of being removed.

These changes have been tested in singleplayer and multiplayer to be working.

There is 1 minor issue however. The system before was easier to unreserve a vehicle if the player left a squad. However, there is no hook for leaving SL or ASL. So if you reserve a supply truck as SL or ASL and then lose SL or ASL, the spawn will fail, nothing will happen and it just says "You will spawn as x in 0:00 seconds". I think a good idea to catch situations like this is to have a function on the client that is run if the spawn is 0:00 and stuck.