Some fixes/improvements:  * Made it so if there is ever a situation where the player cannot spawn for some (future) reason, the player is alerted on the HUD to select new settings (as a fail safe)  * Now if the player loses SL or ASL while reserving a Logi truck and the spawn fails, it invalidates the user's spawn settings (SpawnPointIndex + VehiclePoolIndex), and opens the player's deploy menu  * Now if a vehicle fails to spawn, it removes the reservation (as it couldn't be spawned)
Notes:
Pretty sure the system should have been removing reservation for spawn failure. However other checks were in place to make it so it wasn't needed, but it should have been there in the first place. I did not create a hook to unreserve the vehicle as soon as the player loses ASL or SL, because that is sorta a hack anyways. Plus its not really needed at this point, if the player waits for the spawn timer and then it fails all they have to do is select valid spawn settings and they can spawn right away.