Some changes to constructions:  * Increased health on all 3 sandbags (roughly double)  * Increased health on dragoons tooth and lowered damage multiplier from HE splash
Notes:
Sandbags are near useless, as 1 tank can roll in and unload an HE round to destroy everything. Now it will take quite a few HE rounds to destroy a lot of fortifications. This will increase their value and hopefully we'll see more use out of them. Anyways should be interesting to see how this change affects things.