Added: DH_Vegetation2_stc.Bushes.Bushes_Row_01
Added a row of bushes that is commonly used in several WIP maps, but fixed the origin so this static can be used in future levels (levelers cannot replace the old one with this fixed one because of the origin fix). This is why you should never use a static with an offset, its makes it so you can't automatically replace it for another static/actor, instead you have to spend much time manually replacing.