Weapons can now have recoil based on recent number of shots fired. This effectively makes it so we can tweak weapons to perform better with short bursts or even medium bursts. This opens the door to making weapons a bit more authentic.
How it works:
In a weapon's fire class, you set a default property (a curve). The curve will look like this:
`RecoilCurve=(Points=((InVal=0.0,OutVal=0.33),(InVal=3.0,OutVal=0.66),(InVal=6.0,OutVal=1.33),(InVal=15.0,OutVal=1.0),(InVal=10000000000.0,OutVal=1.0)))`

The input value is RecoilGain, which is a value which increases for each shot fired, but decreases with time. More specifically each time you fire, it reduces RecoilGain based on how long ago you fired. This we'll call the FallOff rate. The current FallOff rate and Gain Rate is WIP!!! The output value for the curve is the recoil modifier.

From my understanding the current values require a weapon to have at least 300 rpm to be functional with this system. Before doing too many guns we have to decide on the FallOff and Gain rates, because once we start doing more weapon's curves, changing those values will break our curves and require us to re-tweak them all.

Notes:
Again this is highly WIP and will need tweaking. I will commit the Thompson right after this to demonstrate this system. This commit does not actually change any weapon, as the default curve is always a 1.0 modifier.