Added a setting in HUD, which makes it more obvious how to disable chat messages.
Notes:
Its directly linked to the # of messages (nu_MsgCount), meaning if you have 0, its the same as turning off chat messages. This is not intuitive however, hence why I've added this setting.

I put in logic to update the # of messages based on changes on the ShowChat boolean and vice versa.