Initial Commit of Patron Foxholes + Marking Enemy Targets:  * Added a new command menu which will appear for a patron, if they are not a SL or ASL, but are in a Squad  * Patrons can now place foxholes via their command menu  * Patrons can now place enemy markers via their command menu OR on the map via context menu  * Setup system to only show constructions on the radial list that you are ALLOWED to place  * Setup system to allow specific constructions to be placeable by Patrons (or any other factor) with a function override  * Added a IsPatron() function to PRI for quick check if a PRI is a patron or not
Notes:
Thanks to Basnett, because he designed & coded this amazing system that was easy to understand and modify to support this endeavor. Tested to work in MP, tested to make sure it did not negatively affect SLs or ASLs abilities. Further testing is wanted with multiple players. Again thanks Basnett, this system is amazing!