Added WIP recoil changes to the MG42
Notes:
So the idea is to double the base recoil, but reduce recoil when firing smaller bursts (so its roughly the same recoil when firing short bursts).

Long bursts receive worse recoil than base, but it levels out to base after very long bursts.

From my understanding it was doctrine (trained) to fire the MG42 in short bursts. This wasn't because of recoil per say, but more for ammo conserving AND target acquisition. That said, we effectively make using short bursts beneficial to both ammo conservation and target acquisition in DH with this change. As you can fire more accurately with shorter bursts (using up less ammo) and of course target acquisition comes naturally.

This however is WIP and shouldn't be considered finalized settings.