DP28 changes:
* Increased the accuracy significantly (this weapon was known to be very accurate for a machine gun)
* Increased the base recoil
* Removed spread reduction modifiers for crouch/prone/bipod (this makes the weapon have base spread when hip-firing and other stances)
* Increased recoil when deployed
* Added a recoil curve which values bursts of 6 or less
* Added a needed reduction for recoil while hip-firing (it was near useless before, this should make it more viable)

Notes:
This is sorta a buff to the DP28, a likely needed one as I never really properly tweaked it from RO values. However, where it was buffed in accuracy and hip-fire, it now suffers with sustained fire. (must use short controlled bursts or else heavy recoil kicks in). From what I've tested so far, this gun is much much more fun to use.