BAR Changes:  * Increased sway, mostly effects iron-sight standing (rifle mode). But this was a big and heavy weapon. It should have more sway and now does.  * Increased accuracy significantly (spread lowered from 130 to 80)  * Increased recoil % when prone/deployed  * Raised base recoil values  * Added recoil curve which promotes 3-round-bursts (less recoil when RecoilGain is <=3)
Notes:
This makes the BAR a bit harder to use as a rifle, but more dangerous as one in the arms of a skilled soldier. A good player will like the accuracy buff and learn to deal with the additional sway and recoil. As a LMG, its about as effective as it was before. However, it really promotes 3-round-bursts, longer bursts will lose control quickly. These changes should significantly raise the skill cap on the BAR. I think good players will appreciate this change, but moderate or low-skilled players will find these changes a bit troublesome. I'd really have to use this in combat several times to get a better picture of where these changes take the BAR.